backend_fastapi/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
Pillow
python-dotenv
requests
